+++ src/utils/payload.rs
@@ -1,9 +1,10 @@
+use super::crypto::sha256hash;
+use crate::utils::{filter::FilterPayload, get_current_time_nanos};
 use eyre::Result;
 use fastbloom_rs::BloomFilter;
+use libsecp256k1::SecretKey;
 use serde::{Deserialize, Serialize};
 use uuid::Uuid;
-
-use crate::utils::{filter::FilterPayload, get_current_time_nanos};
 
 /// A computation task is the task of computing a result from a given input. The result is encrypted with the public key of the requester.
 /// Plain result is signed by the compute node's private key, and a commitment is computed from the signature and plain result.
@@ -15,15 +16,47 @@ use crate::utils::{filter::FilterPayload, get_current_time_nanos};
 pub struct TaskResponsePayload {
     /// The unique identifier of the task.
     pub task_id: String,
-    /// A signature on the digest of plaintext result, prepended with task id.
+    /// Signature of the payload with task id, hexadecimally encoded.
     pub signature: String,
-    /// Computation result encrypted with the public key of the task.
+    /// Result encrypted with the public key of the task, Hexadecimally encoded.
     pub ciphertext: String,
 }
 
 impl TaskResponsePayload {
     pub fn to_string(&self) -> Result<String> {
         serde_json::to_string(&serde_json::json!(self)).map_err(Into::into)
+    }
+
+    /// Creates the payload of a computation result.
+    ///
+    /// - Sign `task_id || payload` with node `self.secret_key`
+    /// - Encrypt `result` with `task_public_key`
+    pub fn new(
+        payload: impl AsRef<[u8]>,
+        task_id: &str,
+        encrypting_public_key: &[u8],
+        signing_secret_key: &SecretKey,
+    ) -> Result<Self> {
+        // create the message `task_id || payload`
+        let mut preimage = Vec::new();
+        preimage.extend_from_slice(task_id.as_ref());
+        preimage.extend_from_slice(payload.as_ref());
+
+        // sign the message
+        // TODO: use `sign_recoverable` here instead?
+        let digest = libsecp256k1::Message::parse(&sha256hash(preimage));
+        let (signature, recid) = libsecp256k1::sign(&digest, signing_secret_key);
+        let signature: [u8; 64] = signature.serialize();
+        let recid: [u8; 1] = [recid.serialize()];
+
+        // encrypt payload itself
+        let ciphertext = ecies::encrypt(encrypting_public_key, payload.as_ref())?;
+
+        Ok(TaskResponsePayload {
+            ciphertext: hex::encode(ciphertext),
+            signature: format!("{}{}", hex::encode(signature), hex::encode(recid)),
+            task_id: task_id.to_string(),
+        })
     }
 }
 
@@ -54,12 +87,4 @@ impl<T> TaskRequestPayload<T> {
             public_key: public_key.unwrap_or_default(),
         }
     }
-}
-
-/// A parsed `TaskRequestPayload`.
-#[derive(Debug, Clone)]
-pub struct TaskRequest<T> {
-    pub task_id: String,
-    pub(crate) input: T,
-    pub(crate) public_key: Vec<u8>,
 }
